requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyyaml>=6.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
